refactor to add hot or not fields into separate data store
