+++ src/canister/individual_user_template/src/api/hot_or_not_bet/get_hot_or_not_bet_details_for_this_post.rs
@@ -3,13 +3,18 @@ use std::time::SystemTime;
 
 use candid::Principal;
 use shared_utils::{
-    canister_specific::individual_user_template::types::hot_or_not::BettingStatus,
+    canister_specific::individual_user_template::types::{cents, hot_or_not::BettingStatus},
     common::utils::system_time::{self},
 };
 
 use crate::{
     api::canister_management::update_last_access_time::update_last_canister_functionality_access_time,
-    data_model::CanisterData, CANISTER_DATA,
+    data_model::{
+        cents_hot_or_not_game::{self, CentsHotOrNotGame},
+        pump_n_dump::TokenBetGame,
+        CanisterData,
+    },
+    CANISTER_DATA, PUMP_N_DUMP,
 };
 
 #[deprecated]
@@ -33,13 +38,19 @@ fn get_hot_or_not_bet_details_for_this_post_v1(post_id: u64) -> BettingStatus {
     let request_maker = ic_cdk::caller();
     update_last_canister_functionality_access_time();
 
-    CANISTER_DATA.with(|canister_data_ref_cell| {
-        get_hot_or_not_bet_details_for_this_post_impl_v1(
-            &canister_data_ref_cell.borrow(),
-            &system_time::get_current_system_time_from_ic(),
-            &request_maker,
-            post_id,
-        )
+    PUMP_N_DUMP.with_borrow_mut(|token_bet_game| {
+        CANISTER_DATA.with_borrow_mut(|canister_data| {
+            let cents_hot_or_not_game = CentsHotOrNotGame {
+                canister_data,
+                token_bet_game,
+            };
+            get_hot_or_not_bet_details_for_this_post_impl_v1(
+                cents_hot_or_not_game,
+                &system_time::get_current_system_time_from_ic(),
+                &request_maker,
+                post_id,
+            )
+        })
     })
 }
 
@@ -62,19 +73,31 @@ fn get_hot_or_not_bet_details_for_this_post_impl(
 }
 
 fn get_hot_or_not_bet_details_for_this_post_impl_v1(
-    canister_data: &CanisterData,
+    cents_hot_or_not_game: CentsHotOrNotGame,
     current_time: &SystemTime,
     request_maker: &Principal,
     post_id: u64,
 ) -> BettingStatus {
-    let post = canister_data.get_post(&post_id).unwrap();
+    let post = cents_hot_or_not_game
+        .canister_data
+        .get_post(&post_id)
+        .unwrap();
 
     post.get_hot_or_not_betting_status_for_this_post_v1(
         current_time,
         request_maker,
-        &canister_data.hot_or_not_bet_details.room_details_map,
-        &canister_data.hot_or_not_bet_details.post_principal_map,
-        &canister_data.hot_or_not_bet_details.slot_details_map,
+        &cents_hot_or_not_game
+            .token_bet_game
+            .hot_or_not_bet_details
+            .room_details_map,
+        &cents_hot_or_not_game
+            .token_bet_game
+            .hot_or_not_bet_details
+            .post_principal_map,
+        &cents_hot_or_not_game
+            .token_bet_game
+            .hot_or_not_bet_details
+            .slot_details_map,
     )
 }
