+++ src/canister/individual_user_template/src/api/hot_or_not_bet/get_hot_or_not_bets_placed_by_this_profile_with_pagination.rs
@@ -3,7 +3,7 @@ use shared_utils::canister_specific::individual_user_template::types::hot_or_not
 
 use crate::{
     api::canister_management::update_last_access_time::update_last_canister_functionality_access_time,
-    CANISTER_DATA,
+    CANISTER_DATA, PUMP_N_DUMP,
 };
 
 const PAGINATION_PAGE_SIZE: usize = 10;
@@ -31,9 +31,8 @@ fn get_hot_or_not_bets_placed_by_this_profile_with_pagination_v1(
     last_index_sent: usize,
 ) -> Vec<PlacedBetDetail> {
     update_last_canister_functionality_access_time();
-    CANISTER_DATA.with(|canister_data_ref_cell| {
-        canister_data_ref_cell
-            .borrow()
+    PUMP_N_DUMP.with_borrow(|token_bet_game| {
+        token_bet_game
             .hot_or_not_bet_details
             .all_hot_or_not_bets_placed
             .iter()
