+++ src/canister/individual_user_template/src/api/hot_or_not_bet/get_individual_hot_or_not_bet_placed_by_this_profile.rs
@@ -8,7 +8,7 @@ use shared_utils::{
 
 use crate::{
     api::canister_management::update_last_access_time::update_last_canister_functionality_access_time,
-    CANISTER_DATA,
+    CANISTER_DATA, PUMP_N_DUMP,
 };
 
 #[deprecated]
@@ -33,9 +33,8 @@ fn get_individual_hot_or_not_bet_placed_by_this_profile_v1(
     post_id: PostId,
 ) -> Option<PlacedBetDetail> {
     update_last_canister_functionality_access_time();
-    CANISTER_DATA.with(|canister_data_ref_cell| {
-        canister_data_ref_cell
-            .borrow()
+    PUMP_N_DUMP.with_borrow(|token_bet_game| {
+        token_bet_game
             .hot_or_not_bet_details
             .all_hot_or_not_bets_placed
             .get(&(canister_id, post_id))
