+++ src/canister/individual_user_template/src/api/token/get_rewarded_for_referral.rs
@@ -62,7 +62,7 @@ fn get_rewarded_for_referral(referrer: Principal, referree: Principal) {
 
     PUMP_N_DUMP.with_borrow_mut(|pump_and_dump| {
         let referral_reward = pump_and_dump.referral_reward.clone();
-        pump_and_dump.handle_token_event(TokenEvent::Mint {
+        pump_and_dump.cents.handle_token_event(TokenEvent::Mint {
             amount: referral_reward.0.try_into().unwrap(),
             details: MintEvent::Referral {
                 referee_user_principal_id: referrer,
