+++ src/canister/individual_user_template/src/lib.rs
@@ -6,7 +6,7 @@ use api::{
     profile::update_profile_display_details::UpdateProfileDetailsError,
 };
 use candid::{Nat, Principal};
-use data_model::pump_n_dump::PumpAndDumpGame;
+use data_model::pump_n_dump::{PumpAndDumpGame, TokenBetGame};
 use data_model::CanisterData;
 use ic_cdk::api::management_canister::provisional::CanisterId;
 use ic_cdk_macros::export_candid;
@@ -60,7 +60,7 @@ mod util;
 thread_local! {
     static CANISTER_DATA: RefCell<CanisterData> = RefCell::default();
     static SNAPSHOT_DATA: RefCell<Vec<u8>> = RefCell::default();
-    static PUMP_N_DUMP: RefCell<PumpAndDumpGame> = RefCell::default();
+    static PUMP_N_DUMP: RefCell<TokenBetGame> = RefCell::default();
 }
 
 export_candid!();
