+++ src/lib/shared_utils/src/canister_specific/individual_user_template/types/mod.rs
@@ -1,5 +1,8 @@
 pub mod arg;
+pub mod cdao;
+pub mod cents;
 pub mod configuration;
+pub mod device_id;
 pub mod error;
 pub mod follow;
 pub mod hot_or_not;
@@ -8,8 +11,6 @@ pub mod migration;
 pub mod ml_data;
 pub mod post;
 pub mod profile;
+pub mod pump_n_dump;
 pub mod session;
 pub mod token;
-pub mod cdao;
-pub mod device_id;
-pub mod pump_n_dump;
